+++ ci/appveyor_run_tests.ps1
@@ -11,7 +11,12 @@ if ($FALSE -and $env:PLATFORM -eq "x86"){
 # Executable paths for OpenCover
 # Note if OpenCover fails, it won't affect the exit codes.
 $OPENCOVER = Resolve-Path .\packages\OpenCover.*\tools\OpenCover.Console.exe
-$CS_RUNNER = Resolve-Path .\packages\NUnit.*\tools\"$CS_RUNNER".exe
+if ($env:BUILD_OPTS -eq "--xplat"){
+    $CS_RUNNER = Resolve-Path $env:USERPROFILE\.nuget\packages\nunit.consolerunner\*\tools\"$CS_RUNNER".exe
+}
+else{
+    $CS_RUNNER = Resolve-Path .\packages\NUnit.*\tools\"$CS_RUNNER".exe
+}
 $PY = Get-Command python
 
 # Can't use ".\build\*\Python.EmbeddingTest.dll". Missing framework files.
@@ -37,6 +42,23 @@ Write-Host ("Starting embedded tests") -ForegroundColor "Green"
 $CS_STATUS = $LastExitCode
 if ($CS_STATUS -ne 0) {
     Write-Host "Embedded tests failed" -ForegroundColor "Red"
+}
+
+if ($env:BUILD_OPTS -eq "--xplat"){
+    if ($env:PLATFORM -eq "x64") {
+         $DOTNET_CMD = "dotnet"
+    }
+    else{
+         $DOTNET_CMD = "c:\Program Files (x86)\dotnet\dotnet"
+    }
+
+    # Run Embedded tests for netcoreapp2.0 (OpenCover currently does not supports dotnet core)
+    Write-Host ("Starting embedded tests for netcoreapp2.0") -ForegroundColor "Green"
+    &$DOTNET_CMD .\src\embed_tests\bin\netcoreapp2.0_publish\Python.EmbeddingTest.dll
+    $CS_STATUS = $LastExitCode
+    if ($CS_STATUS -ne 0) {
+        Write-Host "Embedded tests for netcoreapp2.0 failed" -ForegroundColor "Red"
+    }
 }
 
 # Set exit code to fail if either Python or Embedded tests failed