+++ src/runtime/pythonexception.cs
@@ -13,6 +13,7 @@ namespace Python.Runtime
         private IntPtr _pyTB = IntPtr.Zero;
         private string _tb = "";
         private string _message = "";
+        private string _pythonTypeName = "";
         private bool disposed = false;
 
         public PythonException()
@@ -32,6 +33,8 @@ namespace Python.Runtime
                 {
                     type = pyTypeName.ToString();
                 }
+
+                _pythonTypeName = type;
 
                 Runtime.XIncref(_pyValue);
                 using (var pyValue = new PyObject(_pyValue))
@@ -57,6 +60,10 @@ namespace Python.Runtime
 
         ~PythonException()
         {
+            // We needs to disable Finalizers until it's valid implementation.
+            // Current implementation can produce low probability floating bugs.
+            return;
+
             Dispose();
         }
 
@@ -128,6 +135,13 @@ namespace Python.Runtime
             get { return _tb; }
         }
 
+        /// <summary>
+        /// Python error type name.
+        /// </summary>
+        public string PythonTypeName
+        {
+            get { return _pythonTypeName; }
+        }
 
         /// <summary>
         /// Dispose Method