+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 -   Fixed conversion of 'float' and 'double' values (#486)
 -   Fixed 'clrmethod' for python 2 (#492)
 -   Fixed double calling of constructor when deriving from .NET class (#495)
+-   Fixed `clr.GetClrType` when iterating over `System` members (#607) 
 
 
 ## [2.3.0][] - 2017-03-11